refactoring, FormPane encapsulates form methods
